# ParkingManagerCoordinateService/src/main/webapp/WEB-INF/swagger.json
{
  "swagger": "2.0",
...
  "paths": {
...
    "/parking/coordinates": {
      "get": {
        "tags": [
          "parking"
        ],
        "summary": "returns dummy coordinates",
        "description": "",
        "operationId": "getCoordinates",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      },
      "post": {
        "tags": [
          "parking"
        ],
        "summary": "creates the coordinates",
        "description": "",
        "operationId": "createCoordinates",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/Coordinates"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Coordinates": {
      "type": "object",
      "required": [
        "latitude",
        "longitude"
      ],
      "properties": {
        "city": {
          "type": "string"
        },
        "longitude": {
          "type": "number",
          "minimum": -180.00,
          "maximum": 180.00,
          "exclusiveMinimum": false,
          "exclusiveMaximum": false
        },
        "latitude": {
          "type": "number",
          "minimum": -180.00,
          "maximum": 180.00,
          "exclusiveMinimum": false,
          "exclusiveMaximum": false
        }
      }
    }
  }
}